简易的登陆注册

+++ nodejs-server-test/nodejs-test-sign-in-demo/server.js
@@ -21,21 +21,44 @@ var server = http.createServer(function(request, response){
 
   console.log('方方说：含查询字符串的路径\n' + pathWithQuery)
 
-  // if(path === '/'){
-  //   response.statusCode = 200
-  //   response.setHeader('Content-Type', 'text/html;charset=utf-8')
-  //   response.write('哈哈哈')
-  //   response.end()
-  // }else{
-  //   response.statusCode = 404
-  //   response.setHeader('Content-Type', 'text/html;charset=utf-8')
-  //   response.write('呜呜呜')
-  //   response.end()
-  // }
-
-
   if(path === '/'){
     var string = fs.readFileSync('./index.html', 'utf8')
+    let cookies
+    if (request.headers.cookie !== undefined) {
+      cookies = request.headers.cookie.split(';')
+    } else {
+      cookies = [ 'sign_in_email=不知道' ]
+    }
+    console.log(cookies);
+    let hash = {}
+    for (let i = 0; i < cookies.length; i++) {
+      let parts = cookies[i].split('=')
+      let key = parts[0]
+      let value = parts[1]
+      hash[key] = value
+    }
+    let email = hash.sign_in_email
+    let users = fs.readFileSync('./db/users', 'utf8')
+    users = JSON.parse(users)
+    for (let i = 0; i < users.length; i++) {
+      if (users[i].email === email) {
+        var foundUser = users[i]
+        break
+      }
+    }
+    console.log('foundUser',foundUser);
+    if (foundUser) {
+      string = string.replace('__password__', foundUser.password)
+    } else {
+      string = string.replace('__password__', '不知道')
+    }
+
+    try {
+      users = JSON.parse(users)
+    } catch (exception) {
+      users = []
+    }
+    console.log('request.headers.cookie', request.headers.cookie);
     response.statusCode = 200
     response.setHeader('Content-Type', 'text/html;charset=utf-8')
     response.write(string)
@@ -68,10 +91,7 @@ var server = http.createServer(function(request, response){
         let value = parts[1]
         hash[key] = decodeURIComponent(value) //转码
       })
-      console.log(hash);
-      // let email = hash['email']
-      // let password = hash['password']
-      // let password_confirmation = hash['password_confirmation']
+
       let {email, password, password_confirmation} = hash
       if(email.indexOf('@') === -1){
         response.statusCode = 400
@@ -81,40 +101,92 @@ var server = http.createServer(function(request, response){
             "email": "invalid"
           }
         }`)
-      } else if(password!== password_confirmation) {
+      } else if(password !== password_confirmation) {
         response.statusCode = 400
+        response.setHeader('Content-Type', 'application/json;charset=utf-8')
         response.write('password not match')
-      } else {
+      } else if(password === password_confirmation){
         var users = fs.readFileSync('./db/users', 'utf8')
-        console.log('users', users);
+
         try {
           users = JSON.parse(users)
         } catch (exception) {
           users = []
         }
         let inUse = false
+        console.log('users', users);
         for (let i = 0; i < users.length; i++) {
-          let user[i] = users[i]
+          let user = users[i]
           if (user.email === email) {
-            inUse = ture
+            inUse = true
             break
           }
         }
         if (inUse) {
+          console.log('---------');
           response.statusCode = 404
-          response.setHeader('Content-Type', 'application/jacascript;charset=utf-8')
-          response.write('呜呜呜')
+          response.setHeader('Content-Type', 'application/json;charset=utf-8')
+          response.write(`{
+            "errors": {
+              "users": "ture"
+            }
+          }`)
           response.end()
+          return
+        } else {
+          users.push({email: email, password: password})
+          var usersString = JSON.stringify(users)
+          console.log('usersString', usersString);
+          fs.writeFileSync('./db/users', usersString)
         }
-        users.push({email: email, password: password})
-        var usersString = JSON.stringify(users)
-        console.log('usersString', usersString);
-        fs.writeFileSync('./db/users', usersString)
-
         response.statusCode = 200
       }
       response.end()
     })
+  } else if(path === '/sign_in' && method === 'GET') {
+    let string = fs.readFileSync('./sign_in.html', 'utf8')
+    response.statusCode = 200
+    response.setHeader('Content-Type', 'text/html;charset=utf-8')
+    response.write(string)
+    response.end()
+  } else if(path === '/sign_in' && method === 'POST'){
+    readBody(request).then( (body)=>{
+      let strings = body.split('&')
+      let hash = []
+      strings.forEach( (string)=>{
+        let parts = string.split('=')
+        let key =  parts[0]
+        let value = parts[1]
+        hash[key] = decodeURIComponent(value) //转码
+      })
+
+      let {email, password} = hash
+      console.log(email);
+      console.log(password);
+      var users = fs.readFileSync('./db/users', 'utf8')
+
+      try {
+        users = JSON.parse(users)
+      } catch (exception) {
+        users = []
+      }
+      let found
+      for (let i = 0; i < users.length; i++) {
+        if (users[i].email === email && users[i].password === password) {
+          found = true
+          break
+        }
+      }
+      if(found){
+        //Set-Cookie: <cookie-name>=<cookie-value>
+        response.setHeader(`Set-Cookie`, `sign_in_email=${email}`)
+        response.statusCode = 200
+      } else {
+        response.statusCode = 401
+      }
+
+      response.end()
+      })
   } else {
     response.statusCode = 404
     response.setHeader('Content-Type', 'application/jacascript;charset=utf-8')
